requirements.txt
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
colander==1.8.3
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
deltalake==0.6.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
google-analytics-data==0.16.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.5.0
iso8601==1.1.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.31
proto-plus==1.22.2
protobuf==4.21.12
psutil==5.9.2
pure-eval==0.2.2
py==1.11.0
py4j==0.10.9.5
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
Pygments==2.13.0
pyodbc==4.0.35
pyparsing==3.0.9
pyspark==3.3.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pywin32==304
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SpeechRecognition==3.8.1
stack-data==0.5.1
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
translationstring==1.4
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.11.0
alembic==1.8.1
astroid==2.12.5
attrs==22.1.0
autopep8==1.7.0
azure-common==1.1.28
azure-core==1.25.1
azure-functions==1.11.2
azure-servicebus==7.8.0
black==22.6.0
cachetools==5.2.1
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colander==1.8.3
coverage==6.4.4
deltalake==0.6.4
dill==0.3.5.1
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.8.0
flake8==5.0.4
google-ads==19.0.0
google-analytics-data==0.14.2
google-api-core==2.10.1
google-auth==2.16.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
greenlet==1.1.3
grpcio==1.51.1
grpcio-status==1.51.1
identify==2.5.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflect==6.0.0
iniconfig==1.1.1
iso8601==1.0.2
isodate==0.6.1
isort==5.10.1
lazy-object-proxy==1.7.1
Mako==1.2.2
MarkupSafe==2.1.1
mccabe==0.7.0
msrest==0.7.1
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.2
oauthlib==3.2.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
proto-plus==1.22.1
protobuf==4.21.12
py==1.11.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
pylint==2.15.0
pymssql==2.2.5
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.40
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
translationstring==1.4
typing_extensions==4.3.0
uamqp==1.6.0
urllib3==1.26.12
virtualenv==20.16.3
wrapt==1.14.1
xlrd==2.0.1
zipp==3.8.1